Moves around types, updates rule data
