+++ src/sentry/notifications/notification_action/types.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import logging
 from abc import ABC, abstractmethod
 from collections.abc import Callable, Collection, Sequence
