+++ src/sentry/notifications/platform/types.py
@@ -7,10 +7,7 @@ from typing import Any, Literal, Protocol
 
 from pydantic import BaseModel
 
-from pydantic import BaseModel, ConfigDict
-
 from sentry.integrations.types import ExternalProviderEnum
-from sentry.models.rule import Rule
 
 
 class NotificationCategory(StrEnum):
@@ -99,29 +96,6 @@ NOTIFICATION_SOURCE_MAP: dict[NotificationCategory, list[NotificationSource]] = 
         NotificationSource.SEER_EXPLORER_ERROR,
     ],
 }
-
-
-class NotificationRuleInfo(BaseModel):
-    """
-    A plain-data representation of a rule that triggered a notification.
-    Used instead of the Django Rule ORM model to decouple the notification
-    platform from the ORM layer.
-    """
-
-    model_config = ConfigDict(frozen=True)
-
-    id: int
-    label: str
-    data: dict[str, Any]
-    environment_id: int | None = None
-
-    def to_rule(self) -> Rule:
-        return Rule(
-            id=self.id,
-            label=self.label,
-            data=self.data,
-            environment_id=self.environment_id,
-        )
 
 
 class NotificationProviderKey(StrEnum):
